requirements.txt
Flask==2.0.2
Flask-RESTful==0.3.9
tensorflow==2.13.0
tensorflow-intel==2.13.0
keras==2.13.1
numpy==1.26.4
pandas==2.0.3
Pillow==10.1.0
scikit-learn==1.3.2
opencv-python==4.8.1.78
matplotlib==3.7.4
requests==2.31.0
